@@ -1,143 +1,105 @@
-﻿using OpenTK;
-using OpenTK.Audio.OpenAL;
+﻿using OpenTK.Audio.OpenAL;
 using RPVoiceChat.Utils;
-using System;
 
 namespace RPVoiceChat.Audio
 {
     public static class OALW
     {
-        private static ContextHandle context;
-        private static ContextHandle _activeContext;
-        private static object audio_context_lock = new object();
-
-        public static void InitContext()
-        {
-            lock (audio_context_lock)
-            {
-                if (context != null) Alc.DestroyContext(context);
-
-                var device = Alc.OpenDevice(null);
-                var attrs = new int[0];
-                context = Alc.CreateContext(device, attrs);
-            }
-        }
-
-        public static void ExecuteInContext(Action action)
-        {
-            if (context == null) throw new Exception("OpenAL audio context has not been initialized.");
-            var ctx = SetCurrentContext();
-
-            action();
-
-            ResetContext(ctx);
-        }
-
-        public static T ExecuteInContext<T>(Func<T> function)
-        {
-            if (context == null) throw new Exception("OpenAL audio context has not been initialized.");
-            var ctx = SetCurrentContext();
-
-            var result = function();
-
-            ResetContext(ctx);
-            return result;
-        }
-
         public static void Source(int source, ALSourceb property, bool value)
         {
-            ExecuteInContext(() =>
-            {
-                AL.Source(source, property, value);
-                CheckError($"Error setting source {property}");
-            });
+            AL.Source(source, property, value);
+            CheckError($"Error setting source {property}");
         }
 
         public static void Source(int source, ALSourcef property, float value)
         {
-            ExecuteInContext(() =>
-            {
-                AL.Source(source, property, value);
-                CheckError($"Error setting source {property}");
-            });
+            AL.Source(source, property, value);
+            CheckError($"Error setting source {property}");
+        }
+
+        public static void Source(int source, ALSource3f property, float value1, float value2, float value3)
+        {
+            AL.Source(source, property, value1, value2, value3);
+            CheckError($"Error setting source {property}");
         }
 
         public static void Listener(ALListenerf property, float value)
         {
-            ExecuteInContext(() =>
-            {
-                AL.Listener(property, value);
-                CheckError($"Error setting listener's {property}");
-            });
+            AL.Listener(property, value);
+            CheckError($"Error setting listener {property}");
         }
 
         public static int GenSource()
         {
-            return ExecuteInContext(() =>
-            {
-                var source = AL.GenSource();
-                CheckError("Error generating source");
+            var source = AL.GenSource();
+            CheckError("Error generating source");
 
-                return source;
-            });
+            return source;
+        }
+
+        public static void GetSource(int source, ALGetSourcei property, out int value)
+        {
+            AL.GetSource(source, property, out value);
+            CheckError($"Error getting source {property}");
         }
 
         public static ALSourceState GetSourceState(int source)
         {
-            return ExecuteInContext(() =>
-            {
-                var state = AL.GetSourceState(source);
-                CheckError("Error getting source state");
+            GetSource(source, ALGetSourcei.SourceState, out var state);
 
-                return state;
-            });
+            return (ALSourceState)state;
+        }
+
+        public static void SourceQueueBuffer(int source, int buffer)
+        {
+            AL.SourceQueueBuffer(source, buffer);
+            CheckError("Error SourceQueueBuffer");
+        }
+
+        public static int SourceUnqueueBuffer(int source)
+        {
+            var buffer = AL.SourceUnqueueBuffer(source);
+            CheckError("Error SourceUnqueueBuffer", ALError.InvalidValue);
+
+            return buffer;
         }
 
         public static void SourcePlay(int source)
         {
-            ExecuteInContext(() =>
-            {
-                AL.SourcePlay(source);
-                CheckError("Error playing source");
-            });
+            AL.SourcePlay(source);
+            CheckError("Error playing source");
         }
 
         public static void SourceStop(int source)
         {
-            ExecuteInContext(() =>
-            {
-                AL.SourceStop(source);
-                CheckError("Error stop playing source");
-            });
+            AL.SourceStop(source);
+            CheckError("Error stop playing source");
         }
 
         public static void DeleteSource(int source)
         {
-            ExecuteInContext(() =>
-            {
-                AL.DeleteSource(source);
-                CheckError("Error deleting source");
-            });
+            AL.DeleteSource(source);
+            CheckError("Error deleting source");
         }
 
         public static int[] GenBuffers(int bufferCount)
         {
-            return ExecuteInContext(() =>
-            {
-                var buffers = AL.GenBuffers(bufferCount);
-                CheckError("Error gen buffers");
+            var buffers = AL.GenBuffers(bufferCount);
+            CheckError("Error gen buffers");
 
-                return buffers;
-            });
+            return buffers;
+        }
+
+        public static void BufferData(int currentBuffer, ALFormat format, byte[] audio, int frequency)
+        {
+            AL.BufferData(currentBuffer, format, audio, frequency);
+            CheckError("Error buffer data");
         }
 
         public static void DeleteBuffers(int[] buffers)
         {
-            ExecuteInContext(() =>
-            {
-                AL.DeleteBuffers(buffers);
-                CheckError("Error deleting buffers");
-            });
+            AL.DeleteBuffers(buffers);
+            CheckError("Error deleting buffers");
         }
 
         public static void ClearError()
@@ -151,38 +113,7 @@ namespace RPVoiceChat.Audio
             if (error == ALError.NoError) return;
             if (ignoredErrors == error) return;
 
-            if (_activeContext != context)
-            {
-                Logger.client.Debug($"Calling CheckError outside of mod context is incorrect. Marker of incorrect call: \"{Value}\"");
-                return;
-            }
-
             Logger.client.Debug("{0} {1}", Value, AL.GetErrorString(error));
-        }
-
-        private static ContextHandle? SetCurrentContext()
-        {
-            lock (audio_context_lock)
-            {
-                var oldContext = Alc.GetCurrentContext();
-                if (oldContext == context) return null;
-
-                Alc.MakeContextCurrent(context);
-                _activeContext = context;
-
-                return oldContext;
-            }
-        }
-
-        private static void ResetContext(ContextHandle? ctx)
-        {
-            lock (audio_context_lock)
-            {
-                if (ctx == null) return;
-
-                Alc.MakeContextCurrent((ContextHandle)ctx);
-                _activeContext = (ContextHandle)ctx;
-            }
         }
     }
 }
